Fix upsertCredentials
`'Performing an update on the path '_id' would modify the immutable field '_id''` could occur in `configurePaymentCredentialsModule.ts`
(e.g., during the Datatrans test). MongoDB does not support upserts where the inserted ID (in the setOnInsert)
is different then the queried one (https://jira.mongodb.org/browse/SERVER-18076).
This occured when an `_id` was provided, no document was found and the insert was provided with a newly generated ID. This can no longer happen.