Check if current user is admin is in acl but the check if impersonated userId belongs to an admin is in mutation
The rational behind this is that we want to have an absolutely clear „return true“ for admins and an absolutely clear „return false“ for other users, that is the most important check and needs to be in ACL and easily auditable. It would also allow to easily just extend and allow other users to do impersonation but by default only admin’s should be able to impersonate at all and only should they be able to impersonate other users that are not admin.

I had to change this because the way it was done was allowing any user to do impersonation (not properly tested) as longs as it was not an admin.